[I25-273] feat : 프로젝트 Play 버튼 기능 추가 및 사이드바 컴포넌트 모듈화
