+++ src/app/components/project/sidebar/ProjectTechnologiesSection.tsx
@@ -1,0 +1,20 @@
+import { Label } from '@/app/components/system/text';
+import SkillTagProvider from '@/app/components/modal/inputs/SkillTagProvider';
+
+interface ProjectTechnologiesSectionProps {
+  technologies: number[];
+}
+
+export const ProjectTechnologiesSection = ({
+  technologies,
+}: ProjectTechnologiesSectionProps) => (
+  <section className="flex-col gap-[0.375rem]">
+    <Label>기술스택</Label>
+    {technologies.length > 0 ? (
+      <SkillTagProvider readOnly initialTags={technologies} />
+    ) : (
+      <Label className="text-detail !block">등록된 정보가 없습니다.</Label>
+    )}
+  </section>
+);
+
